requirements.txt
notebook
pysqlite3-binary
langchain
langchain-community
langchain-groq
unstructured
selenium
chromadb
streamlit
pandas
python-dotenv